refactored code, split app.js into separate modules/components
