--- image-gallery/app/public/data/schnoodles.json
@@ -1,30 +1,0 @@
-[{ 
-  "title": "Breezy",
-  "description": "Breezy saying hi",
-  "url": "/img/breezy-odemen.JPG",
-  "id": 0
-},
-{ 
-  "title": "Two naughty dogs",
-  "description": "Breezy and Princess at the dog park",
-  "url": "/img/dog-park.JPG",
-  "id": 1
-},
-{ 
-  "title": "Cutie pie",
-  "description": "Princess just lounging",
-  "url": "/img/princess-cat-bed.JPG",
-  "id": 2
-},
-{ 
-  "title": "Cuddling",
-  "description": "Sharing the dog bed",
-  "url": "/img/schnoodles-cat-bed.JPG",
-  "id": 3
-},
-{ 
-  "title": "Merry Christmas",
-  "description": "Princess wearing her Santa suite",
-  "url": "/img/princess-santa.JPG",
-  "id": 4
-}]
